fix stepper props
