+++ website/src/routes/schedules/$scheduleId/form/-components/StepperData.tsx
@@ -1,0 +1,10 @@
+type Steps = {
+  title: string;
+};
+export const steps: Steps[] = [
+  { title: "座席・チケット選択" },
+  { title: "購入情報入力" },
+  { title: "購入内容確認" },
+  { title: "お支払い" },
+  { title: "購入完了" },
+];
