+++ website/src/routes/schedules/$scheduleId/form/-components/result/Reserved.tsx
@@ -1,24 +1,21 @@
+import { Stepper } from "@yamada-ui/react";
 import styled from "styled-components";
-import Stepper from "./-components/Stepper";
-import BuyContent from "./-components/BuyContent";
-import CustomerInfo from "./-components/CustomerInfo";
-import PaymentInfo from "./-components/PaymentInfo";
+import BuyContent from "./BuyContent";
+import CustomerInfo from "./CustomerInfo";
+import PaymentInfo from "./PaymentInfo";
 
 export function Index() {
   return (
     <ReservedContainer>
-      <Stepper />
+      <Stepper index={0} />
       <Title>
         <h1>予約内容確認</h1>
         <p>
           以下の内容を確認の上、「決済する」ボタンを押して決済を完了してください。
         </p>
       </Title>
-
       <BuyContent />
-
       <CustomerInfo />
-
       <PaymentInfo />
     </ReservedContainer>
   );
